Physics update
I decided that for simplicitys sake I will use raycasts that will be going down till they reach the ground and that will be the deciding factor of where the ball  will be rolling. I have a front back left and right ray cast that checks the direction in which the ball is suppoused to be moving thowards. Its moving in the direction of the longest raycast.